output: initial implementation of netcdf writing
WIP: needs CRS work, CF work, axis ordering inspection, and testing

This change switches from numpy array output to netcdf files.  The
domain information is not written, as it is a data structure and not a
value.

A flag is added for the output file, and the output directory remains
for the plotting output if requested.